requirements.txt
Django 
djangorestframework  
psycopg2  
drf-spectacular    

 djangorestframework-simplejwt 
 djangorestframework-simplejwt[crypto]
 oauth2
 django-oauth-toolkit  
 django-allauth
# mkdocs
 